Attempts to fix broken test.

+++ LXSupportKitTests/LXSupportKitTests.m
@@ -68,7 +68,7 @@
                    LXAdjustToScale(LXCGFloor, 0.8, [[UIScreen mainScreen] scale]));
     XCTAssertEqual(LXAdjustToScale(LXCGFloor, 0.8, 1.0), 0.0);
     XCTAssertEqual(LXAdjustToScale(LXCGFloor, 0.8, 2.0), 0.5);
-    XCTAssertEqual(LXAdjustToScale(LXCGFloor, 0.8, 3.0), 2.0 / 3.0);
+    XCTAssertEqual(LXAdjustToScale(LXCGFloor, 0.8, 3.0), (CGFloat)2.0 / (CGFloat)3.0);
 }
 
 - (void)lx_assertOnMainThread {
